Safari-609.1.11.1/LayoutTests/http/tests/security/srcdoc-in-sandbox-cannot-access-parent-expected.txt

CONSOLE MESSAGE: line 4: SecurityError: Sandbox access violation: Blocked a frame at "null" from accessing a cross-origin frame.  The frame requesting access is sandboxed and lacks the "allow-same-origin" flag.
PASS
